requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pydantic==2.*
orjson
